+++ Samples/SentrySampleShared/SentrySampleUITestShared/SentrySampleUITestShared.h
@@ -1,9 +1,0 @@
-#import <Foundation/Foundation.h>
-
-//! Project version number for SentrySampleUITestShared.
-FOUNDATION_EXPORT double SentrySampleUITestSharedVersionNumber;
-
-//! Project version string for SentrySampleUITestShared.
-FOUNDATION_EXPORT const unsigned char SentrySampleUITestSharedVersionString[];
-
-#import <SentrySampleUITestShared/SentryProcessInfo.h>