+++ Samples/SentrySampleShared/Sources/SentrySampleShared/LoremIpsumText.swift
@@ -2,10 +2,10 @@ import Foundation
 
 public class BundleResourceProvider: NSObject {
     public static var loremIpsumTextFilePath: String? {
-        Bundle(for: self).path(forResource: "LoremIpsum", ofType: "txt")
+        Bundle.module.path(forResource: "LoremIpsum", ofType: "txt")
     }
 
     @objc public static var screenshotURL: URL? {
-        Bundle(for: self).url(forResource: "screenshot", withExtension: "png")
+        Bundle.module.url(forResource: "screenshot", withExtension: "png")
     }
 }