+++ Samples/iOS-Swift/App/Sources/TransactionsViewController.swift
@@ -1,5 +1,5 @@
-import Sentry
 import SentrySampleShared
+import SentrySwift
 import UIKit
 
 // swiftlint:disable force_unwrapping private_outlet